requirements.txt
beautifulsoup4==4.13.4
Pillow==11.2.1
Requests==2.32.4
streamlit==1.45.1
streamlit_option_menu==0.4.0
